farmer model added, stack animation added

+++ Assets/Scripts/Game/Component/Stat/PlayerStats.cs
@@ -3,8 +3,8 @@
     public struct PlayerStats
     {
         public float MaxSpeed;
-        public float MaxCapacity;
-        public float Capacity;
+        public int MaxCapacity;
+        public int Capacity;
        
 
     }
